requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.11
langchain-openai==0.2.14
python-dotenv==1.0.1
pydantic==2.10.4
